chore: move teamsBot code from mustache to template and modify eslint config
