+++ templates/bot/js/.eslintrc
@@ -1,8 +1,6 @@
 {
+    "parser": "@typescript-eslint/parser",
     "parserOptions": {
-        "ecmaVersion": 2017
-    },
-    "env":{
-        "es6": true
+        "sourceType": "module"
     }
 }
