+++ templates/bot/ts/default/teamsBot.ts
@@ -1,24 +1,31 @@
-{{#IS_ME}}
-const axios = require("axios");
-const querystring = require("querystring");
-{{/IS_ME}}
-const { TeamsActivityHandler, CardFactory{{#IS_BOT}}, TurnContext{{/IS_BOT}}} = require("botbuilder");
-{{#IS_BOT}}
-const rawWelcomeCard = require("./adaptiveCards/welcome.json");
-const rawLearnCard = require("./adaptiveCards/learn.json");
-const cardTools = require("@microsoft/adaptivecards-tools");
-{{/IS_BOT}}
 
-class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} extends TeamsActivityHandler {
-  {{#IS_BOT}}
+import { default as axios } from "axios";
+import * as querystring from "querystring";
+
+import { TeamsActivityHandler, CardFactory, TurnContext, AdaptiveCardInvokeValue, AdaptiveCardInvokeResponse} from "botbuilder";
+
+import rawWelcomeCard from "./adaptiveCards/welcome.json"
+import rawLearnCard from "./adaptiveCards/learn.json"
+import { AdaptiveCards } from "@microsoft/adaptivecards-tools";
+
+export interface DataInterface {
+  likeCount: number
+}
+
+
+export class TeamsBotMessageExtensionBot extends TeamsActivityHandler {
+  
+  // record the likeCount
+  likeCountObj: { likeCount: number };
+
   constructor() {
     super();
 
-    // record the likeCount
     this.likeCountObj = { likeCount: 0 };
 
     this.onMessage(async (context, next) => {
       console.log("Running with Message Activity.");
+
       let txt = context.activity.text;
       const removedMentionText = TurnContext.removeRecipientMention(
         context.activity
@@ -31,13 +38,13 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
       // Trigger command by IM text
       switch (txt) {
         case "welcome": {
-          const card = cardTools.AdaptiveCards.declareWithoutData(rawWelcomeCard).render();
+          const card = AdaptiveCards.declareWithoutData(rawWelcomeCard).render();
           await context.sendActivity({ attachments: [CardFactory.adaptiveCard(card)] });
           break;
         }
         case "learn": {
           this.likeCountObj.likeCount = 0;
-          const card = cardTools.AdaptiveCards.declare(rawLearnCard).render(this.likeCountObj);
+          const card = AdaptiveCards.declare<DataInterface>(rawLearnCard).render(this.likeCountObj);
           await context.sendActivity({ attachments: [CardFactory.adaptiveCard(card)] });
           break;
         }
@@ -53,12 +60,11 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
       await next();
     });
 
-    // Listen to MembersAdded event, view https://docs.microsoft.com/en-us/microsoftteams/platform/resources/bot-v3/bots-notifications for more events
     this.onMembersAdded(async (context, next) => {
       const membersAdded = context.activity.membersAdded;
       for (let cnt = 0; cnt < membersAdded.length; cnt++) {
         if (membersAdded[cnt].id) {
-          const card = cardTools.AdaptiveCards.declareWithoutData(rawWelcomeCard).render();
+          const card = AdaptiveCards.declareWithoutData(rawWelcomeCard).render();
           await context.sendActivity({ attachments: [CardFactory.adaptiveCard(card)] });
           break;
         }
@@ -69,25 +75,31 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
 
   // Invoked when an action is taken on an Adaptive Card. The Adaptive Card sends an event to the Bot and this
   // method handles that event.
-  async onAdaptiveCardInvoke(context, invokeValue) {
+  async onAdaptiveCardInvoke(
+    context: TurnContext,
+    invokeValue: AdaptiveCardInvokeValue
+  ): Promise<AdaptiveCardInvokeResponse> {
     // The verb "userlike" is sent from the Adaptive Card defined in adaptiveCards/learn.json
     if (invokeValue.action.verb === "userlike") {
       this.likeCountObj.likeCount++;
-      const card = cardTools.AdaptiveCards.declare(rawLearnCard).render(this.likeCountObj);
+      const card = AdaptiveCards.declare<DataInterface>(rawLearnCard).render(this.likeCountObj);
       await context.updateActivity({
         type: "message",
         id: context.activity.replyToId,
         attachments: [CardFactory.adaptiveCard(card)],
       });
-      return { statusCode: 200 };
+      return { statusCode: 200, type: undefined, value: undefined };
     }
   }
-  {{/IS_BOT}}
+  
 
-  {{#IS_ME}}
+  
   // Messaging extension Code
   // Action.
-  handleTeamsMessagingExtensionSubmitAction(context, action) {
+  public async handleTeamsMessagingExtensionSubmitAction(
+    context: TurnContext,
+    action: any
+  ): Promise<any> {
     switch (action.commandId) {
       case "createCard":
         return createCardCommand(context, action);
@@ -99,7 +111,7 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
   }
 
   // Search.
-  async handleTeamsMessagingExtensionQuery(context, query) {
+  public async handleTeamsMessagingExtensionQuery(context: TurnContext, query: any): Promise<any> {
     const searchQuery = query.parameters[0].value;
     const response = await axios.get(
       `http://registry.npmjs.com/-/v1/search?${querystring.stringify({
@@ -129,7 +141,10 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
     };
   }
 
-  async handleTeamsMessagingExtensionSelectItem(context, obj) {
+  public async handleTeamsMessagingExtensionSelectItem(
+    context: TurnContext,
+    obj: any
+  ): Promise<any> {
     return {
       composeExtension: {
         type: "result",
@@ -140,8 +155,8 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
   }
 
   // Link Unfurling.
-  handleTeamsAppBasedLinkQuery(context, query) {
-    const attachment = CardFactory.thumbnailCard("Thumbnail Card", query.url, [query.url]);
+  public async handleTeamsAppBasedLinkQuery(context: TurnContext, query: any): Promise<any> {
+    const attachment = CardFactory.thumbnailCard("Image Preview Card", query.url, [query.url]);
 
     const result = {
       attachmentLayout: "list",
@@ -154,11 +169,11 @@ class {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} ex
     };
     return response;
   }
-  {{/IS_ME}}
+  
 }
 
-{{#IS_ME}}
-function createCardCommand(context, action) {
+
+async function createCardCommand(context: TurnContext, action: any): Promise<any> {
   // The user has chosen to create a card by choosing the 'Create Card' context menu command.
   const data = action.data;
   const heroCard = CardFactory.heroCard(data.title, data.text);
@@ -178,7 +193,7 @@ function createCardCommand(context, action) {
   };
 }
 
-function shareMessageCommand(context, action) {
+async function shareMessageCommand(context: TurnContext, action: any): Promise<any> {
   // The user has chosen to share a message by choosing the 'Share Message' context menu command.
   let userName = "unknown";
   if (
@@ -229,6 +244,3 @@ function shareMessageCommand(context, action) {
     },
   };
 }
-{{/IS_ME}}
-
-module.exports.{{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}} = {{#IS_BOT}}TeamsBot{{/IS_BOT}}{{^IS_BOT}}MessageExtensionBot{{/IS_BOT}};
